rename all to files_mindmap2 & update node & change project_dir
Updated the namespace to Files_MindMap2 and id to files_mindmap2 trough the whole codebase.
Updated the package with node.
Added some exclude Files for the Release in Makefile
Changed the project_dir= variable, as when run in the same folder it does not work. but not sure how the workflow for nextcloud is. (this change might be wrong for that)